+++ app/models/work_package.rb
@@ -924,12 +924,29 @@ class WorkPackage < ActiveRecord::Base
   end
 
   def add_time_entry_for(user, attributes)
-    return if attributes.nil? || attributes.values.all?(&:blank?)
+    return if time_entry_blank?(attributes)
 
     attributes.reverse_merge!({ :user => user,
                                 :spent_on => Date.today })
 
     time_entries.build(attributes)
+  end
+
+  ##
+  # Checks if the time entry defined by the given attributes is blank.
+  # A time entry counts as blank despite a selected activity if that activity
+  # is simply the default activity and all other attributes are blank.
+  def time_entry_blank?(attributes)
+    return true if attributes.nil?
+    key = "activity_id"
+    id = attributes[key]
+    default_id = if id && !id.blank?
+      Enumeration.exists? :id => id, :is_default => true, :type => 'TimeEntryActivity'
+    else
+      true
+    end
+
+    default_id && attributes.except(key).values.all?(&:blank?)
   end
 
   # >>> issues.rb >>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>>