+++ doc/CHANGELOG.md
@@ -34,6 +34,7 @@ See doc/COPYRIGHT.rdoc for more details.
 * `#2594` Fix: [Activity] Too many filter selects than necessary
 * `#3332` [CodeClimate] Mass Assignment AuthSourcesController
 * `#3333` [CodeClimate] Mass Assignment RolesController
+* `#3438` Activity default value makes log time required
 * `#3481` Fix: [Activity] Not possible to unselect all filters
 * `#3854` Move function and Query filters allows to select groups as responsible
 * `#3974` [Timelines] Typo at creating timelines